requirements.txt
asgiref==3.4.1
Django==3.2.9
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
ffmpeg==1.4
importlib-metadata==4.8.2
Markdown==3.3.6
netifaces==0.11.0
protobuf==3.17.1
pydub==0.25.1
PyJWT==2.3.0
python-nmap==0.7.1
pytz==2021.3
pywatchman==1.4.1
six==1.16.0
sqlparse==0.4.2
zipp==3.6.0
